Add configuration option to replace argLine property.
This adds a new configuration option, "propertyOverride" which allows
full replacement of the argLine property. The current behaviour (keeping the
old value) is not desirable if the property is not actually argLine itself
but some other property which holds a default value. This change allows using

  <properties>
    <myProperty>default-value</myProperty>
  </properties>

...

  <plugin>
     ... jacoco plugin ...
    <configuration>
        <propertyName>myProperty</propertyName>
        <propertyOverride>true</propertyOverride>
    </configuration>
  </plugin>

and have the value be either the default-value (if the prepare-agent
goal is not run) or the agent line (if the prepare-agent goal is run)
instead of the agent line with the default-value appended.