Make FORMAT_VERSION non-constant
If a constant expression is assigned to a static final variable, javac
will inline the constant value in the caller, which means constants
should only be used for values that will never change.  Otherwise,
calling code must be recompiled when upgrading to a new version of the
library even if the caller would otherwise not require any changes.