Add filter for Kotlin lateinit properties
This change adds a new filter for Kotlin lateinit properties by
searching in the bytecode for the following sequence of instructions:

    IFNONNULL ...
    LDC "..."
    INVOKESTATIC kotlin/jvm/internal/Intrinsics.throwUninitializedPropertyAccessException (Ljava/lang/String;)V

and ignoring this sequence if found. The downside of this approach is
that every instruction in every method must be scanned. Perhaps this
filter could only be limited to Kotlin classes.