Change assert for CLI DumpTest to pass on non-English systems
For a non-English system the error message may differ.

E.g. on a Hungarian system it says:
"[WARN] Kapcsolat elutasítva (Connection refused)."
This fails the test as it does not include "[WARN] Connection refused".

Change the assert to allow for the changed error message.
Also include an additional assert to ensure the message is a WARN.