Use branch index 1 of the jumps into the case instead of 0.
Currently KotlinWhenStringFilter breaks branch coverage for String
based when statements because the replacement branches into the case
statements can never be covered.

This is because the new branches point to the "first" branch of the
unconditional jumps to the bodies of the case statement (after the
String equality check).

But the first branch of a jump is the "fallthrough" (the path if the
jump is not taken), which is never taken for a simple goto. The
target of a jump is the second branch.

Fixes https://github.com/jacoco/jacoco/issues/1938